requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
crontab==0.23.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-static==0.0.6
Django==3.2
django-admin-interface==0.16.1
django-colorfield==0.4.1
django-crispy-forms==1.11.2
django-crontab==0.7.1
django-extensions==3.1.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-freezer==0.1.4
django-import-export==2.5.0
django-static-fontawesome==5.14.0.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
fontawesome-free==5.15.3
gunicorn==20.1.0
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
PyYAML==5.4.1
sentry-sdk==1.0.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.3.1
static3==0.7.0
tablib==3.0.0
urllib3==1.26.4
xlrd==2.0.1
xlwt==1.3.0
